Rework PopupMenu and Table components
