+++ app/src/lib/tokens/PopupMenu.svelte
@@ -1,0 +1,9 @@
+
+<div class="dropdown dropdown-hover dropdown-top">
+	<div class="overflow-x-auto dropdown-content z-[1] menu shadow-xl bg-info text-info-content rounded-box w-96 max-h-80">
+		<slot name="popup_content" />
+	</div>
+	<div role="button">
+		<slot name="button_content" />
+	</div>
+</div>
